Editing Implemented
To Do:
Add border (white letters, black borders) to text and move it to the bottom of the screen.
Exception/Error handling (bug that says the file is in use when trying to delete it, might have to terminate the process or delete the variables after usage.
Integrate Youtube Uploading
potentially integrate reddit comments (on the bottom after the Title, Poster text).
Server side implementation (hosting/whatever)
Speed Up Rendering if possible